+++ app/src/main/AndroidManifest.xml
@@ -25,7 +25,7 @@
             android:value="${MAPS_API_KEY}" />
 
         <activity
-            android:name="com.hearhere.presentation.login.SplashActivity"
+            android:name="com.hearhere.presentation.features.login.SplashActivity"
                   android:exported="true"
             android:screenOrientation="portrait">
