+++ presentation/src/main/java/com/hearhere/presentation/features/login/LoginViewModel.kt
@@ -1,4 +1,4 @@
-package com.hearhere.presentation.login
+package com.hearhere.presentation.features.login
 
 import android.app.Application
 import android.util.Log
