+++ presentation/src/main/java/com/hearhere/presentation/features/login/RoutingViewModel.kt
@@ -1,4 +1,4 @@
-package com.hearhere.presentation.login
+package com.hearhere.presentation.features.login
 
 import android.util.Log
 import androidx.lifecycle.viewModelScope
