+++ presentation/src/main/java/com/hearhere/presentation/features/login/SetNickNameActivity.kt
@@ -1,4 +1,4 @@
-package com.hearhere.presentation.login
+package com.hearhere.presentation.features.login
 
 import android.os.Bundle
 import android.util.Log
