+++ presentation/src/main/java/com/hearhere/presentation/features/post/PostSelectEmotionFragment.kt
@@ -1,4 +1,4 @@
-package com.hearhere.presentation.post
+package com.hearhere.presentation.features.post
 
 import android.util.Log
 import androidx.fragment.app.activityViewModels
