+++ presentation/src/main/java/com/hearhere/presentation/features/post/PostSelectWeatherFragment.kt
@@ -1,4 +1,4 @@
-package com.hearhere.presentation.post
+package com.hearhere.presentation.features.post
 
 import android.os.Bundle
 import androidx.fragment.app.activityViewModels
