+++ presentation/src/main/java/com/hearhere/presentation/features/post/SearchSingerFragment.kt
@@ -1,4 +1,4 @@
-package com.hearhere.presentation.post
+package com.hearhere.presentation.features.post
 
 import android.os.Bundle
 import android.util.Log
