+++ presentation/src/main/res/layout/activity_post.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="viewModel"
-            type="com.hearhere.presentation.post.PostViewModel" />
+            type="com.hearhere.presentation.features.post.PostViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
