+++ presentation/src/main/res/layout/activity_post_select_option.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="viewModel"
-            type="com.hearhere.presentation.post.PostSelectOptionViewModel" />
+            type="com.hearhere.presentation.features.post.PostSelectOptionViewModel" />
     </data>
 
     <ScrollView
