+++ presentation/src/main/res/layout/fragment_search_singer.xml
@@ -7,13 +7,13 @@
 
         <variable
             name="viewmodel"
-            type="com.hearhere.presentation.post.PostViewModel" />
+            type="com.hearhere.presentation.features.post.PostViewModel" />
     </data>
 
     <FrameLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".post.SearchSingerFragment">
+        tools:context=".features.post.SearchSingerFragment">
 
         <ScrollView
