+++ presentation/src/main/res/layout/item_music_list.xml
@@ -3,7 +3,7 @@
     <data>
         <variable
             name="binder"
-            type="com.hearhere.presentation.post.adapter.MusicListItemBinder" />
+            type="com.hearhere.presentation.features.post.adapter.MusicListItemBinder" />
     </data>
     <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
         android:layout_width="match_parent"
